# PCD Performance test harness