--- a/packages/tools/perftest/.eslintrc.js
+++ b/packages/tools/perftest/.eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extends: ["@pcd/eslint-config-custom"],
+  root: true,
+};