--- a/packages/tools/perftest/.gitignore
+++ b/packages/tools/perftest/.gitignore
@@ -1,0 +1,4 @@
+*.js
+*.ts.map
+!.eslintrc.js
+dist