--- a/packages/tools/perftest/CHANGELOG.md
+++ b/packages/tools/perftest/CHANGELOG.md
@@ -1,0 +1,7 @@
+# @pcd/perftest
+
+## 0.0.1
+
+### Patch Changes
+
+- Initial changeset.