--- a/packages/tools/perftest/package.json
+++ b/packages/tools/perftest/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "commander": "^11.0.0",
     "conf": "^11.0.2",
+    "log-symbols": "^4.1.0",
     "uuid": "^9.0.0",
     "@pcd/eddsa-ticket-pcd": "^0.5.0",
     "@pcd/pcd-types": "^0.10.0",