--- a/packages/tools/perftest/package.json
+++ b/packages/tools/perftest/package.json
@@ -16,6 +16,7 @@
     "@pcd/eddsa-ticket-pcd": "^0.5.0",
     "@pcd/pcd-types": "^0.10.0",
     "@pcd/semaphore-identity-pcd": "^0.10.0",
+    "@pcd/semaphore-signature-pcd": "^0.10.0",
     "@pcd/util": "^0.4.0",
     "@pcd/zk-eddsa-event-ticket-pcd": "^0.4.0",
     "@semaphore-protocol/identity": "^3.14.0"