--- a/packages/tools/perftest/src/cases/EdDSATicketTimer.ts
+++ b/packages/tools/perftest/src/cases/EdDSATicketTimer.ts
@@ -10,7 +10,7 @@ import { v4 as uuid } from "uuid";
 import { TimerCase } from "../types";
 
 // Key borrowed from https://github.com/iden3/circomlibjs/blob/4f094c5be05c1f0210924a3ab204d8fd8da69f49/test/eddsa.js#L103
-export const prvKey: string =
+export const prvKey =
   "0001020304050607080900010203040506070809000102030405060708090001";
 
 export const ticketData: ITicketData = {