--- a/packages/tools/perftest/src/index.ts
+++ b/packages/tools/perftest/src/index.ts
@@ -1,0 +1,17 @@
+#!/usr/bin/env node
+
+import { program } from "commander";
+import logSymbols from "log-symbols";
+import "./timer";
+
+program
+  .name("pcd-perftest")
+  .description("Command line tool for testing PCD performance.")
+  .version("0.0.1", "-v, --version", "Show PCD pertest CLI version.")
+  .configureOutput({
+    outputError: (message) => {
+      console.info(`\n ${logSymbols.error}`, message);
+    }
+  });
+
+program.parse(process.argv);