--- a/packages/tools/perftest/tsconfig.json
+++ b/packages/tools/perftest/tsconfig.json
@@ -1,0 +1,36 @@
+{
+  "extends": "@pcd/tsconfig/ts-library.json",
+  "compilerOptions": {
+    "lib": [
+      "dom"
+    ],
+    "outDir": "dist",
+    "declarationDir": "dist/types",
+    // Include all source files, including tests
+    "rootDir": "."
+  },
+  // Some artifact-including packages need to import a JSON file
+  "include": [
+    "src",
+    "test",
+    "artifacts/*.json"
+  ],
+  "exclude": [
+    "dist",
+    "build",
+    "node_modules"
+  ],
+  // DO NOT MODIFY MANUALLY BEYOND THIS POINT
+  // References are automatically maintained by `yarn fix-references`
+  "references": [
+    {
+      "path": "../../pcd/eddsa-ticket-pcd"
+    },
+    {
+      "path": "../../lib/pcd-types"
+    },
+    {
+      "path": "../../lib/util"
+    }
+  ]
+}