--- a/packages/tools/perftest/tsconfig.json
+++ b/packages/tools/perftest/tsconfig.json
@@ -33,6 +33,9 @@
       "path": "../../pcd/semaphore-identity-pcd"
     },
     {
+      "path": "../../pcd/semaphore-signature-pcd"
+    },
+    {
       "path": "../../lib/util"
     },
     {