--- a/packages/tools/perftest/typedoc.json
+++ b/packages/tools/perftest/typedoc.json
@@ -1,0 +1,3 @@
+{
+  "entryPoints": ["src/index.ts"]
+}