Adds a command-line tool for rough local perf testing of PCD operations, by running them repeatedly and averaging.  It has cases built for eddsa-ticket-pcd and zk-eddsa-event-ticket-pcd and room to easily add more.  My intent is to use this to verify any changes to cryptographic libraries don't regress our performance too badly, starting with a retroactive test of #1493 (see comment on that PR).

Can be run like this:

    cd zupass/packages/tools/perftest
    yarn start timer all all

Here's a snippet of example output from testing #1493:

```
Running timing test of zk-eddsa-event-ticket-pcd.prove...
✔ init: 0.15695810317993164ms
✔ setup: 44.70187520980835ms
✔ first op: 1298.1974577903748ms
✔ second op: 519.0421657562256ms
✔ repeat op (average of 10): 514.5606165885926ms
Running timing test of zk-eddsa-event-ticket-pcd.verify...
✔ init: 0.07541704177856445ms
✔ setup: 566.9950408935547ms
✔ first op: 9.173666000366211ms
✔ second op: 7.883166313171387ms
✔ repeat op (average of 10): 8.090016603469849ms
```

I'm sure the TypeScript is a mess, and I'm open to feedback for learning purposes.  Seems to work well enough for what it is, though.